requirements.txt
wget
Pillow
numpy
opencv-python
tweepy
pygame
sklearn
scipy
matplotlib